+++ wx/lib/ogl/diagram.py
