+++ wx/lib/ogl/oglmisc.py
